+++ atlast_sc/instruments/classes/Default.py
@@ -1,6 +1,7 @@
 import astropy.units as u
 from astropy import constants
 from atlast_sc.instrument import Instrument
+from atlast_sc.derived_groups import noise_temperature
 
 """
 Default instrument parameters
@@ -35,7 +36,7 @@ class Default(Instrument):
     # Additional instrument specific methods below #
     ################################################
 
-    def calculate_system_temperature(self, obs_freq, bandwidth, T_cmb, eta_eff, T_atm, 
+    def calculate_system_temperature(self, obs_freq, bandwidth, eta_eff, 
                                      T_amb, T_sky, transmittance, n_pol):
         """
         Returns system temperature, following calculation in [doc]
@@ -47,7 +48,7 @@ class Default(Instrument):
         system_temp = 1 / (eta_eff * transmittance) * \
             (self.T_rx
             + (eta_eff * T_sky)
-            + ((1 - eta_eff) * T_amb)
+            + ((1 - eta_eff) * noise_temperature(T_amb, obs_freq))
             )
         self.T_sys = system_temp
         return system_temp